# Frontend/swagger-openapi.json
{
    "openapi": "3.0.0",
...
    "info": {
        "title": "Identity API",
        "description": "An ASP.NET Core Web API for managing identity items",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Contact",
            "url": "https://example.com/contact"
        },
        "license": {
            "name": "License",
            "url": "https://example.com/license"
        },
        "version": "v1"
    },
    "paths": {
        "/api/auth/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateOrUpdateUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/confirm-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/OTPRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OTPRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OTPRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/OTPRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationTokenObjectResponseApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationTokenObjectResponseApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationTokenObjectResponseApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/reset-email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "emailAddress",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/confirm-reset-email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "emailAddress",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "otpCode",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/Role": {
            "get": {
                "tags": [
                    "Role"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "IsLock",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "in": "query",
                        "name": "Keyword",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "PageSize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "PageNum",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOPaginationItemsApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOPaginationItemsApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOPaginationItemsApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateOrUpdateRoleRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Role"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateOrUpdateRoleRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "guid",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Role/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "operationId": "GetRoleById",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleDTOApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "components": {
        "requestBodies": {
            "CreateOrUpdateUserRequest": {
                "content": {
                    "application/json-patch+json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateUserRequest"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateUserRequest"
                        }
                    },
                    "text/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateUserRequest"
                        }
                    },
                    "application/*+json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateUserRequest"
                        }
                    }
                }
            },
...
            "CreateOrUpdateRoleRequest": {
                "content": {
                    "application/json-patch+json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateRoleRequest"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateRoleRequest"
                        }
                    },
                    "text/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateRoleRequest"
                        }
                    },
                    "application/*+json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateOrUpdateRoleRequest"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\n"
            }
        },
        "schemas": {
...
            "AuthenticationTokenObjectResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "item": {
                        "$ref": "#/components/schemas/AuthenticationToken"
                    }
                },
                "additionalProperties": false
            },
            "AuthenticationTokenObjectResponseApiResponse": {
                "type": "object",
                "properties": {
                    "requestId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "isSuccess": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AuthenticationTokenObjectResponse"
                    }
                },
                "additionalProperties": false
            },
...
            "CreateOrUpdateRoleRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "isLock": {
                        "type": "boolean"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                },
                "additionalProperties": false
            },
...
            "CreateOrUpdateUserRequest": {
                "type": "object",
                "properties": {
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "avatarUrl": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "middleName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneCode": {
                        "type": "string"
                    },
                    "region": {
                        "type": "string"
                    },
                    "isRequiredChangePassword": {
                        "type": "boolean"
                    },
                    "isPasswordExpired": {
                        "type": "boolean"
                    },
                    "expirationDate": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "locked": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "attemptLogin": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "isResentEmail": {
                        "type": "boolean"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssigningService"
                        }
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResponse"
                        }
                    }
                },
                "additionalProperties": false
            },
            "ForgotPasswordRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "password": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "otpCode": {
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "OTPRequest": {
                "type": "object",
                "properties": {
                    "otpCode": {
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
...
            "RoleDTO": {
                "type": "object",
                "properties": {
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    },
                    "isLock": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionDTO"
                        }
                    }
                },
                "additionalProperties": false
            },
            "RoleDTOApiResponse": {
                "type": "object",
                "properties": {
                    "requestId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "isSuccess": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/RoleDTO"
                    }
                },
                "additionalProperties": false
            },
            "RoleDTOPaginationItems": {
                "type": "object",
                "properties": {
                    "pageSize": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "pageNum": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "totalCount": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "totalPage": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleDTO"
                        }
                    }
                },
                "additionalProperties": false
            },
            "RoleDTOPaginationItemsApiResponse": {
                "type": "object",
                "properties": {
                    "requestId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "isSuccess": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/RoleDTOPaginationItems"
                    }
                },
                "additionalProperties": false
            },
...
            "SignInRequest": {
                "type": "object",
                "properties": {
                    "userName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "appKeys": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "additionalProperties": false
            },
...
            "UserDTOApiResponse": {
                "type": "object",
                "properties": {
                    "requestId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "isSuccess": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserDTO"
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}